requirements.txt
flask>=2.0.0
paramiko>=3.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
gunicorn>=21.2.0
beautifulsoup4>=3.1.0
